rules and sort three

+++ libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: jefernan <jefernan@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/02 08:29:07 by jefernan          #+#    #+#             */
-/*   Updated: 2021/09/30 09:17:21 by jefernan         ###   ########.fr       */
+/*   Updated: 2022/06/10 22:08:20 by jefernan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <stdio.h>
 
-int	ft_atoi(const char *str)
+long int	ft_atoi(const char *str)
 {
 	int	i;
-	int	num;
+	long int	num;
 	int	neg;
 
 	i = 0;
